[ADAM-1438] Add ability to save FASTA back as a single file.
Resolves #1438. To support saving FASTA as a single file, moved the
private `writeTextRdd` method from the `FeatureRDD` to `GenomicRDD`, and opened
protections to `protected`.
